requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
black==24.4.2
celery==5.4.0
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
distlib==0.3.8
Django==5.0.7
django-celery-beat==2.6.0
django-debug-toolbar==4.4.6
django-redis==5.4.0
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
Faker==26.0.0
filelock==3.15.4
identify==2.6.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-django==4.8.0
pytest-env==1.1.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
rpds-py==0.19.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
